refactor(provisa-ui): decompose AdminPage into per-tab components
Extract the Local Users, Orgs, Roles, and Observability sections out of the
1600+ line AdminPage into self-contained components under components/admin,
each owning its own state, data fetching, and handlers. AdminPage drops from
1674 to ~540 lines and now renders each tab as a child component.
